requirements.txt
# Core Web Framework
fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.22
pydantic==2.12.5

# LLM & RAG Pipeline
langchain==0.2.14
langchain-core==0.2.43
langchain-community==0.2.12
langchain-groq==0.1.9
groq==0.37.1

# Vector Search & Embeddings
faiss-cpu==1.13.2
sentence-transformers==5.2.2
numpy==1.26.4

# Keyword Search
rank-bm25==0.2.2

# Document Processing
python-docx==1.2.0
pypdf==6.6.2
lxml==6.0.2

# Text Processing & Science
torch==2.10.0
transformers==5.0.0
scikit-learn==1.8.0
scipy==1.17.0
indic-transliteration==2.3.75

# Utilities
python-dotenv==1.2.1
requests==2.32.5
tqdm==4.67.1
tenacity==8.5.0
